requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
openai==0.28.1
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.3